Moving PlanValidator and PlanEnricher to respective folders.
