+++ health-services/plan-service/src/main/java/digit/service/enrichment/PlanEnricher.java
@@ -1,4 +1,4 @@
-package digit.service;
+package digit.service.enrichment;
 
 import digit.web.models.Plan;
 import digit.web.models.PlanRequest;
