+++ health-services/plan-service/src/main/java/digit/service/validator/PlanValidator.java
@@ -1,9 +1,11 @@
-package digit.service;
+package digit.service.validator;
 
 import com.jayway.jsonpath.JsonPath;
 import digit.config.Configuration;
 import digit.repository.PlanConfigurationRepository;
 import digit.repository.PlanRepository;
+import digit.service.PlanEmployeeService;
+import digit.service.enrichment.PlanEnricher;
 import digit.util.BoundaryUtil;
 import digit.util.CampaignUtil;
 import digit.util.CommonUtil;
